method-specific test for to_dict_of_blocks
